requirements.txt
apscheduler==3.6.3
flask==1.1.2
flask-restx==0.2.0
paho-mqtt==1.5.0
pyyaml